chore: move fake classes to utils file
